base14c/requirements.txt
asgiref==3.3.4
Django==3.2.3
django-countries==7.2.1
django-filter==2.4.0
django-phone-field==1.8.1
django-phonenumber-field==5.2.0
mysqlclient==2.0.3
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
xlrd==2.0.1
xlwt==1.3.0
